+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 * Upgrade `geozero` from 0.11 to 0.12.
+* Upgrade `pyo3` from 0.20 to 0.21 and switch to new `Bound` API.
 
 ## 0.7.0 - 2023-10-12
 * Upgrade `pyo3` from 0.19 to 0.20.